Add LiveSwap type and GetLiveSwap trait
Swaps can exist at various stages of their lifecycle, we have tentative
names for two of these stages so far

- 'Created' swap - this is a swap that has been created by a POST call to
the REST API.

- 'Finalized' swap - this is swap after we have shared the data required
  to/from the counterparty node.

The data base module controls these types.  This patch introduces a
third term 'live' swap. The term 'finalized' is reduced to mean just the
additional data shared during the communication protocols.

- 'Live' swap - a swap that has been created, has been 'finalized' (has
  the additional data shared during communication protocols) and is now
  'live' in the system.

RFC: Request For Comments on these names please.

These are used in the following manner:

`CreatedSwap` - store this in the database when a swap is created.
`FinalizedSwap` - store this in the database when a swap is finalized.
`LiveSwap` - retrieve this from the database (how data is read from the
db is the db modules concern, this is the type that the db module
exposes for swap retrieval).